# backend/docs/api-docs.json
{
  "openapi": "3.1.0",
  "info": { "title": "ticket-bus", "description": "", "version": "1.0.0" },
...
  "paths": {
...
    "/vehicles": {
      "get": {
        "summary": "Obtener todos los vehiculos y sus amenities",
        "deprecated": false,
        "description": "",
        "operationId": "getVehicles",
        "tags": ["Vehicles"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Vehiculos obtenidos con éxito!",
                      "body": {
                        "vehicles": [
                          {
                            "id": 1,
                            "number": 1,
                            "plate": "QRST90",
                            "totalSeats": 60,
                            "amenities": [
                              "Cargador USB",
                              "WiFi",
                              "Aire Acondicionado"
                            ]
                          },
                          {
                            "id": 2,
                            "number": 2,
                            "plate": "ABCD12",
                            "totalSeats": 40,
                            "amenities": ["WiFi", "TV"]
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "error",
                      "code": 500,
                      "message": "Error al obtener los vehiculos!"
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      },
      "post": {
        "summary": "Registro de nuevo vehiculo",
        "deprecated": false,
        "description": "",
        "operationId": "createVehicle",
        "tags": ["Vehicles"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Vehicle" },
              "example": {
                "number": 3,
                "plate": "ABCD12",
                "totalSeats": 40,
                "amenities": ["WiFi", "TV"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Vehiculo creado con éxito!",
                      "body": {
                        "vehicle": {
                          "id": 3,
                          "number": 3,
                          "plate": "ABCD12",
                          "totalSeats": 40,
                          "amenities": ["WiFi", "TV"]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflicto en el servidor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "fail",
                      "code": 409,
                      "message": "El vehiculo ya existe."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "error",
                      "code": 500,
                      "message": "No se pudo registrar el vehiculo."
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/vehicles/{id}": {
      "get": {
        "summary": "Obtener un vehiculo por su ID",
        "deprecated": false,
        "description": "",
        "operationId": "getVehicleByID",
        "tags": ["Vehicles"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del vehiculo",
            "required": true,
            "example": "1",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Vehiculo obtenido con éxito!",
                      "body": {
                        "vehicle": {
                          "id": 1,
                          "number": 1,
                          "plate": "QRST90",
                          "totalSeats": 60,
                          "amenities": [
                            "Cargador USB",
                            "WiFi",
                            "Aire Acondicionado"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Registro no encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "ENo se encontró ningún vehiculo con el ID: 1",
                    "value": {
                      "status": "fail",
                      "code": 404,
                      "message": "No se encontró vehiculo con el ID: 0"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "error",
                      "code": 500,
                      "message": "Error al obtener el vehiculo!"
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "Actualizar vehiculo",
        "deprecated": false,
        "description": "",
        "operationId": "updateVehicle",
        "tags": ["Vehicles"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del vehiculo",
            "required": true,
            "example": "1",
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Vehicle" },
              "example": {
                "number": 1,
                "plate": "QRST90",
                "totalSeats": 60,
                "amenities": ["Cargador USB", "WiFi", "Aire Acondicionado"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Vehiculo actualizado con éxito!",
                      "body": {
                        "vehicle": {
                          "id": 1,
                          "number": 1,
                          "plate": "QRST90",
                          "totalSeats": 60,
                          "amenities": [
                            "Cargador USB",
                            "WiFi",
                            "Aire Acondicionado"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Registro no encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "ENo se encontró ningún vehiculo con el ID: 1",
                    "value": {
                      "status": "fail",
                      "code": 404,
                      "message": "No se encontró vehiculo con el ID: 0"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "error",
                      "code": 500,
                      "message": "No se pudo actualizar el vehiculo."
                    }
                  }
                }
              }
            }
          },
          "security": []
        }
      },
      "delete": {
        "summary": "Eliminar vehiculo",
        "deprecated": false,
        "description": "",
        "operationId": "deleteVehicle",
        "tags": ["Vehicles"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del vehiculo",
            "required": true,
            "example": 1,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Vehiculo eliminado con éxito!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Registro no encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "No se encontró el vehiculo",
                    "value": {
                      "status": "fail",
                      "code": 404,
                      "message": "No se encontró vehiculo con el ID: 0"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "error",
                      "code": 500,
                      "message": "No se pudo eliminar el vehiculo."
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/vehicles/plate/{plate}": {
      "get": {
        "summary": "Obtener un vehiculo por su placa",
        "deprecated": false,
        "description": "",
        "operationId": "getVehicleByPlate",
        "tags": ["Vehicles"],
        "parameters": [
          {
            "name": "plate",
            "in": "path",
            "description": "Placa del vehiculo",
            "required": true,
            "example": "QRST90",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Vehiculo obtenido con éxito!",
                      "body": {
                        "vehicle": {
                          "id": 1,
                          "number": 1,
                          "plate": "QRST90",
                          "totalSeats": 60,
                          "amenities": [
                            "Cargador USB",
                            "WiFi",
                            "Aire Acondicionado"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Registro no encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "No se encontró ningún vehiculo con la placa: QRST90",
                    "value": {
                      "status": "fail",
                      "code": 404,
                      "message": "No se encontró vehiculo con la placa: QRST90"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "error",
                      "code": 500,
                      "message": "Error al obtener el vehiculo!"
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "code": { "type": "integer" },
          "message": { "type": "string" },
          "body": { "type": "object", "properties": {} }
        },
        "required": ["status", "code", "message"]
      }
    },
    "securitySchemes": {}
  },
  "servers": [
    { "url": "http://localhost:3000/api/v1", "description": "Develop Env" },
    {
      "url": "https://ticket-bus-api-dev.up.railway.app/api/v1",
      "description": "Prod Env"
    }
  ]
}